+++ CHANGELOG.md
@@ -60,3 +60,4 @@
 - BEIS users (administrators) can mark other users as active or inactive
 - Budget start and end dates are validated according to IATI standard
 - BEIS users can view projects (read-only) and download them as XML via a button
+- Codelist dropdowns do not contain values which have been marked as "status: withdrawn" by IATI